feature: prevent changes in archived rooms

+++ src/websocket/comments.gateway.ts
@@ -1,3 +1,4 @@
+import { BadRequestException } from "@nestjs/common";
 import {
   ConnectedSocket,
   MessageBody,
@@ -6,11 +7,12 @@ import {
 } from "@nestjs/websockets";
 import { plainToInstance } from "class-transformer";
 import { Socket } from "socket.io";
-import { ActivitiesService } from "src/api/activities/activities.service";
-import { UpdateCommentDto } from "src/api/comments/dtos/update-comment.dto";
-import { Comment } from "src/api/comments/entities/comment.entity";
+import { ActivitiesService } from "../api/activities/activities.service";
 import { CommentsService } from "../api/comments/comments.service";
 import { CreateCommentDto } from "../api/comments/dtos/create-comment.dto";
+import { UpdateCommentDto } from "../api/comments/dtos/update-comment.dto";
+import { Comment } from "../api/comments/entities/comment.entity";
+import { RoomsService } from "../api/rooms/rooms.service";
 import { BaseWebsocketGateway } from "./base-websocket.gateway";
 
 @WebSocketGateway()
@@ -18,6 +20,7 @@ export class CommentsGateway extends BaseWebsocketGateway {
   constructor(
     private commentsService: CommentsService,
     private activitiesService: ActivitiesService,
+    private roomsService: RoomsService,
   ) {
     super();
   }
@@ -29,6 +32,10 @@ export class CommentsGateway extends BaseWebsocketGateway {
   ) {
     const { id } = (socket as any).user;
     const { roomId, payload } = data;
+    const room = await this.roomsService.findById(roomId);
+    if (room.isActive === false) {
+      throw new BadRequestException();
+    }
     try {
       const newComment = await this.commentsService.createComment(id, payload);
       this.server.to(roomId).emit("comments/created", plainToInstance(Comment, newComment));
@@ -44,6 +51,10 @@ export class CommentsGateway extends BaseWebsocketGateway {
   ) {
     const { id } = (socket as any).user;
     const { roomId, commentId, payload } = data;
+    const room = await this.roomsService.findById(roomId);
+    if (room.isActive === false) {
+      throw new BadRequestException();
+    }
     try {
       const updatedComment = await this.commentsService.updateComment(id, commentId, payload);
       this.server
@@ -61,6 +72,12 @@ export class CommentsGateway extends BaseWebsocketGateway {
   ) {
     const { id } = (socket as any).user;
     const { roomId, commentId } = data;
+    const room = await this.roomsService.findById(roomId);
+    if (room.isActive === false) {
+      throw new BadRequestException();
+    }
+
+    const comment = await this.commentsService.findById(commentId);
     try {
       const deletedComment = await this.commentsService.deleteComment(commentId);
       this.server.to(roomId).emit("comments/deleted", deletedComment);
